+++ packages/web-pkg/src/configuration/types.ts
@@ -29,6 +29,9 @@ export interface OptionsConfiguration {
   mode?: string
   isRunningOnEos?: boolean
   embedTarget?: string
+  editor?: {
+    openAsPreview?: boolean | string[]
+  }
 }
 
 export interface OAuth2Configuration {