+++ tests/e2e/support/objects/app-files/trashbin/index.ts
@@ -1,4 +1,4 @@
-import { Page } from 'playwright'
+import { Page } from '@playwright/test'
 import * as po from './actions'
 import { SpacesEnvironment } from '../../../environment'
 export class Trashbin {